% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{beta.update}
\alias{beta.update}
\title{Update Beta parameters in a Gibbs Step}
\usage{
beta.update(X, Vinv, y, sigsq.eps, M = 1)
}
\arguments{
\item{X}{Covariate design matrix}

\item{Vinv}{Inverse of V_\{lambda, Z, r\}, calculated with \code{\link{makeVcomps}}}

\item{y}{Outcome vector}

\item{sigsq.eps}{Sigma^2, variance component}

\item{M}{Stochastic approximation constant. If StA is requested, M = K, otherwise M = 1.}
}
\value{
Sampled vector of Betas
}
\description{
See Bobb et al. (2015) supplement, page 4 for details.
}
