% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_binomial}
\alias{bkmr_mcmc_binomial}
\title{Redirect Binomial family using specified link function (default will be 'logit')}
\usage{
bkmr_mcmc_binomial(link, ...)
}
\arguments{
\item{link}{character vector giving link function name. Currently implemented for 'logit' and 'probit'}

\item{...}{All arguments passed through to sampler function}
}
\value{
Output from sampler function
}
\description{
Redirect Binomial family using specified link function (default will be 'logit')
}
