% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_gaussian}
\alias{bkmr_mcmc_gaussian}
\title{Wrapper for bkmr_mcmc_gaussian_oneChain}
\usage{
bkmr_mcmc_gaussian(nchains, warmup, iter, ...)
}
\arguments{
\item{nchains}{Number of MCMC or HMC chains to utilize. Defaults to 1.}

\item{warmup}{Number of iterations to discard as warmups. Defaults to 100}

\item{iter}{number of iterations to run the sampler}

\item{...}{Arguments passed through to bkmr_mcmc_gaussian_oneChain()}
}
\value{
Samples from gaussian MCMC bkmr sampler
}
\description{
Wrapper for bkmr_mcmc_gaussian_oneChain to allow for multiple chains to be run (sequentially, not in parallel)
}
