% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_gaussian_oneChain}
\alias{bkmr_mcmc_gaussian_oneChain}
\title{MCMC code to fit one chain of Gaussian BKMR Model}
\usage{
bkmr_mcmc_gaussian_oneChain(
  y,
  Z,
  X,
  starting.values,
  iter,
  id,
  varsel,
  Znew,
  groups,
  rmethod,
  est.h,
  knots,
  ztest,
  control.params,
  verbose,
  missingX,
  data.comps,
  ...
)
}
\arguments{
\item{y}{a vector of outcome data of length  n.}

\item{Z}{an n-by-M matrix of predictor variables to be included in the h function. Each row represents an observation and each column represents an predictor.}

\item{X}{an n-by-K matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{starting.values}{list of starting values for each parameter. If not specified default values will be chosen. (Only can change if family = gaussian() or binomial(link = 'logit'))}

\item{iter}{number of iterations to run the sampler}

\item{id}{optional vector (of length n) of grouping factors for fitting a model with a random intercept. If NULL then no random intercept will be included.}

\item{varsel}{TRUE or FALSE: indicator for whether to conduct variable selection on the Z variables in h}

\item{Znew}{optional matrix of new predictor values at which to predict h, where each row represents a new observation. This will slow down the model fitting, and can be done as a post-processing step using SamplePred}

\item{groups}{optional vector (of length M) of group indicators for fitting hierarchical variable selection if varsel=TRUE. If varsel=TRUE without group specification, component-wise variable selections will be performed.}

\item{rmethod}{for those predictors being forced into the h function, the method for sampling the r\link{m} values. Takes the value of 'varying' to allow separate r\link{m} for each predictor; 'equal' to force the same r\link{m} for each predictor; or 'fixed' to fix the r\link{m} to their starting values}

\item{est.h}{TRUE or FALSE: indicator for whether to sample from the posterior distribution of the subject-specific effects h_i within the main sampler. This will slow down the model fitting. Only relevant if family = gaussian() or binomial(link = 'logit')}

\item{knots}{optional matrix of knot locations for implementing the Gaussian predictive process of Banerjee et al. (2008). Currently only implemented for models without a random intercept.}

\item{ztest}{optional vector indicating on which variables in Z to conduct variable selection (the remaining variables will be forced into the model).}

\item{control.params}{list of parameters specifying the prior distributions and tuning parameters for the MCMC algorithm. If not specified default values will be chosen. (Only can change if family = gaussian() or binomial(link = 'logit'))}

\item{verbose}{TRUE or FALSE: flag indicating whether to print intermediate diagnostic information during the model fitting.}

\item{missingX}{Logical Flag to indicate if matrix X was provided to the function}

\item{data.comps}{Components needed for V matrix}

\item{...}{Catches extra arguments}
}
\value{
Samples from gaussian MCMC bkmr sampler.
}
\description{
Code slightly modified from original bkmr package, Bobb 2015
}
