% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_logit}
\alias{bkmr_mcmc_logit}
\title{BKMR Sampling on Logit Model using Hamiltonian Monte Carlo}
\usage{
bkmr_mcmc_logit(
  y,
  Z,
  X,
  starting.values = list(rho = 1),
  control.params,
  nchains = 1,
  iter = 1000,
  warmup = 50,
  verbose,
  M = 1,
  ...
)
}
\arguments{
\item{y}{a vector of outcome data of length \code{n}.}

\item{Z}{an \code{n}-by-\code{M} matrix of predictor variables to be included in the \code{h} function. Each row represents an observation and each column represents an predictor.}

\item{X}{an \code{n}-by-\code{K} matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{starting.values}{list of starting values for each parameter. If not specified default values will be chosen. (Only can change if family = gaussian() or binomial(link = 'logit'))}

\item{control.params}{list of parameters specifying the prior distributions and tuning parameters for the MCMC algorithm. If not specified default values will be chosen. (Only can change if family = gaussian() or binomial(link = 'logit'))}

\item{nchains}{Number of MCMC or HMC chains to utilize. Defaults to 1.}

\item{iter}{number of iterations to run the sampler}

\item{warmup}{Number of iterations to discard as warmups. Defaults to 100}

\item{verbose}{TRUE or FALSE: flag indicating whether to print intermediate diagnostic information during the model fitting.}

\item{M}{Stochastic approximation constant. If StA is requested, M = K, otherwise M = 1.}

\item{...}{Catch all to allow package flexibility}
}
\value{
STAN output from bkmr_logit fit
}
\description{
Fits bkmr_logit model using STAN. See file inst/stan/fit_logit.stan for STAN specifications.
}
