% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_poisson}
\alias{bkmr_mcmc_poisson}
\title{BKMR Sampling on Poisson Model using Hamiltonian Monte Carlo}
\usage{
bkmr_mcmc_poisson(
  y,
  Z,
  X,
  starting.values = list(rho = 2),
  nchains = 1,
  iter = 1000,
  warmup = 50,
  ...
)
}
\arguments{
\item{y}{Outcome}

\item{...}{Catch all to allow package flexibility}
}
\value{
STAN output from bkmr_poisson fit
}
\description{
Fits bkmr_poisson model using STAN. See file inst/stan/fit_poisson.stan for STAN specifications.
}
