% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_poisson_log}
\alias{bkmr_mcmc_poisson_log}
\title{BKMR Sampling on Poisson Model using Hamiltonian Monte Carlo}
\usage{
bkmr_mcmc_poisson_log(
  y,
  Z,
  X,
  starting.values = list(rho = 2),
  control.params = list(),
  nchains = 1,
  iter = 1000,
  warmup = 50,
  verbose,
  ...
)
}
\arguments{
\item{y}{a vector of outcome data of length  n.}

\item{Z}{an n-by-M matrix of predictor variables to be included in the h function. Each row represents an observation and each column represents an predictor.}

\item{X}{an n-by-K matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column.}

\item{starting.values}{list of starting values for each parameter. If not specified default values will be chosen. (Only can change if family = gaussian() or binomial(link = 'logit'))}

\item{control.params}{list of parameters specifying the prior distributions and tuning parameters for the MCMC algorithm. If not specified default values will be chosen. (Only can change if family = gaussian() or binomial(link = 'logit'))}

\item{nchains}{Number of MCMC or HMC chains to utilize. Defaults to 1.}

\item{iter}{number of iterations to run the sampler}

\item{warmup}{Number of iterations to discard as warmups. Defaults to 100}

\item{verbose}{TRUE or FALSE: flag indicating whether to print intermediate diagnostic information during the model fitting.}

\item{...}{Catch all to allow package flexibility}
}
\value{
STAN output from bkmr_poisson fit
}
\description{
Fits bkmr_poisson model using STAN. See file inst/stan/fit_poisson.stan for STAN specifications.
}
