% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_probit}
\alias{bkmr_mcmc_probit}
\title{MCMC code to fit Probit BKMR Model}
\usage{
bkmr_mcmc_probit(
  y,
  Z,
  X,
  starting.values,
  nchains,
  iter,
  id,
  warmup,
  control.params,
  ...
)
}
\arguments{
\item{y}{Outcome}
}
\description{
Code slightly modified from original bkmr package, Bobb 2015
}
