% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmrGen-package.R
\docType{package}
\name{bkmrGen-package}
\alias{bkmrGen}
\alias{bkmrGen-package}
\title{bkmrGen: Generalized Bayesian Kernel Machine Regression with Distributed Methods for Computational Scalability}
\description{
Implements Bayesian kernel machine regression (BKMR) under a general link function. In order to improve scalability, MCMC algorithms are run under a distributed divide-and-conquer framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeDuttweiler/bkmrGen}
  \item Report bugs at \url{https://github.com/LukeDuttweiler/bkmrGen/issues}
}

}
\author{
\strong{Maintainer}: Luke Duttweiler \email{lduttweiler@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-0467-995X}{ORCID})

}
\keyword{internal}
