% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{createDataComps}
\alias{createDataComps}
\title{Create Values needed in makeVcomps, depending on if using random intercept model}
\usage{
createDataComps(id, knots)
}
\arguments{
\item{id}{Vector of id values (for random intercept model), default in kmbayes is NULL}

\item{knots}{Vector of knots for Gaussian Process modification of V}
}
\value{
List of options used by makeVcomps to estiamte V
}
\description{
Create Values needed in makeVcomps, depending on if using random intercept model
}
