% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{h.update}
\alias{h.update}
\title{Update estimates of subject-specific health effects in a Gibbs step}
\usage{
h.update(lambda, Vcomps, sigsq.eps, y, X, beta, r, Z, data.comps, M = 1)
}
\arguments{
\item{lambda}{Variance component, tau*sigma^\{-2\}}

\item{Vcomps}{Components of V_\{lambda, Z, r\} matrix, calculated with \code{\link{makeVcomps}}}

\item{sigsq.eps}{Variance component sigma^2}

\item{y}{Outcome vector}

\item{X}{Covariate design matrix}

\item{beta}{Covariate parameter vector}

\item{r}{Augmented variables in kernel matrix}

\item{Z}{Exposure design matrix}

\item{data.comps}{Info for internal VComps.star calculation}

\item{M}{Stochastic approximation constant. If StA is requested, M = K, otherwise M = 1.}
}
\value{
Updated draws of h values
}
\description{
See Bobb et al. (2015) supplement p. 6 for details.
}
