% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{lamAdj}
\alias{lamAdj}
\title{Adjust lambda proposal distribution to improve mixing}
\usage{
lamAdj(lam)
}
\arguments{
\item{lam}{Numeric scalar}
}
\value{
Transformed numeric scalar
}
\description{
This function chooses the mean of the proposal function for the lambda MH step based on the most recent iteration of lambda If this is the identity then E(lambda_star) = lambda_t. Otherwise we have E(lambda_star) = lamAdj(lambda_t).
}
