% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{lambda.update}
\alias{lambda.update}
\title{Update lambda value in Metropolis-Hastings Step}
\usage{
lambda.update(
  r,
  delta,
  lambda,
  whichcomp = 1,
  y,
  X,
  Z = Z,
  beta,
  sigsq.eps,
  Vcomps,
  data.comps,
  control.params,
  M = 1
)
}
\arguments{
\item{r}{Augmented variables in kernel matrix}

\item{delta}{Inclusion parameters for mixture components}

\item{lambda}{Current variance component, tau*sigma^\{-2\}}

\item{whichcomp}{Which component of lambda to operate on (There can be multiple if random intercept model is being used)}

\item{y}{Outcome vector}

\item{X}{Covariate design matrix}

\item{Z}{Exposure design matrix}

\item{beta}{Covariate parameter vector}

\item{sigsq.eps}{Variance component sigma^2}

\item{Vcomps}{Components of V_\{lambda, Z, r\} matrix, calculated with \code{\link{makeVcomps}}}

\item{data.comps}{Info for internal VComps.star calculation}

\item{control.params}{Control parameter list from \code{\link{kmbayes}}, can adjust MH proposal variance or prior values}

\item{M}{Stochastic approximation constant. If StA is requested, M = K, otherwise M = 1.}
}
\value{
MH Step output with updated version of lambda
}
\description{
Update lambda value in Metropolis-Hastings Step
}
