% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{makeKpart}
\alias{makeKpart}
\title{Make K_\{Z,r\} matrix (Exponential kernel)}
\usage{
makeKpart(r, Z1, Z2 = NULL)
}
\arguments{
\item{r}{Vector of augmented variables in kernel matrix controlling smoothness of h()}

\item{Z1}{Exposure design matrix}

\item{Z2}{New exposure design matrix (optional)}
}
\value{
K_\{Z,r\} matrix
}
\description{
See Bobb et al. (2015), supplement page 3 for details.
}
