% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Matrix_to_slam}
\alias{Matrix_to_slam}
\title{Convert 'Matrix' object to 'slam' object}
\usage{
Matrix_to_slam(Mat)
}
\arguments{
\item{Mat}{Object from package 'Matrix'}
}
\value{
Matrix with type 'slam'
}
\description{
Converts a 'Matrix' object to a 'slam' object without reverting to 'matrix' type. This saves a ton of memory and time.
}
