% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{MHstep}
\alias{MHstep}
\title{Process Metropolis-Hastings Step}
\usage{
MHstep(
  r,
  lambda,
  lambda.star,
  r.star,
  delta,
  delta.star,
  y,
  X,
  Z,
  beta,
  sigsq.eps,
  diffpriors,
  negdifflogproposal,
  Vcomps,
  move.type,
  data.comps
)
}
\arguments{
\item{r}{Augmented variables in kernel matrix}

\item{lambda}{Variance component, tau*sigma^\{-2\}}

\item{lambda.star}{Proposal value for lambda}

\item{r.star}{Proposal value for r}

\item{delta}{Inclusion indicator for mixture component}

\item{delta.star}{Proposal value for delta}

\item{y}{Outcome vector}

\item{X}{Covariate design matrix}

\item{Z}{Exposure design matrix}

\item{beta}{Covariate parameter vector}

\item{sigsq.eps}{Variance component sigma^2}

\item{diffpriors}{Difference of priors for M-H calculation}

\item{negdifflogproposal}{Negative difference of the log proposal distributions for MH}

\item{Vcomps}{Components of V_\{lambda, Z, r\} matrix, calculated with \code{\link{makeVcomps}}}

\item{move.type}{Type of move proposed for delta/r, not utilized, just passed through}

\item{data.comps}{Info for interal VComps.star calculation}
}
\value{
List containing new MH values and inputs
}
\description{
Processing function for typical M-H step. Reduces code base by combining likelihood
calculation into one.
}
