% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideConquer.R
\name{mpost_univariate}
\alias{mpost_univariate}
\title{Get Univariate M-Posterior}
\usage{
mpost_univariate(sampList, n_samps, ...)
}
\arguments{
\item{sampList}{List of samples from univariate posterior. Each sample must be a numeric vector.}

\item{n_samps}{Number of samples to return from WASP. Defaults as length of first sub-sample provided in sampList.}

\item{...}{Kept to catch extraneous parameters from other methods.}
}
\value{
n_samps samples from M-Posterior
}
\description{
Takes list of samples from univariate sub-sample posteriors and returns samples from the M-Posterior (Geometric median in the Wasserstein order 1 space). Utilizes Weiszfeld’s Algorithm with Euclidean distance.
}
