% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{newh.update}
\alias{newh.update}
\title{Update estimates for new subject-specific health effects}
\usage{
newh.update(
  Z,
  Znew,
  Vcomps,
  lambda,
  sigsq.eps,
  r,
  y,
  X,
  beta,
  data.comps,
  M = 1
)
}
\arguments{
\item{Z}{Exposure design matrix}

\item{Znew}{Exposure design matrix for new subjects}

\item{Vcomps}{Components of V_\{lambda, Z, r\} matrix, calculated with \code{\link{makeVcomps}}}

\item{lambda}{Variance component, tau*sigma^\{-2\}}

\item{sigsq.eps}{Variance component sigma^2}

\item{r}{Augmented variables in kernel matrix}

\item{y}{Outcome vector}

\item{X}{Covariate design matrix}

\item{beta}{Covariate parameter vector}

\item{data.comps}{Info for internal VComps.star calculation}

\item{M}{Stochastic approximation constant. If StA is requested, M = K, otherwise M = 1.}
}
\value{
Updated samples of new h values
}
\description{
Update estimates for new subject-specific health effects
}
