% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{rdelta.comp.update}
\alias{rdelta.comp.update}
\title{Update r and delta in a component-wise variable selection MCMC step}
\usage{
rdelta.comp.update(
  r,
  delta,
  lambda,
  y,
  X,
  beta,
  sigsq.eps,
  Vcomps,
  Z,
  ztest,
  data.comps,
  control.params,
  rprop.gen2,
  rprop.logdens1,
  rprior.logdens,
  rprior.logdens2,
  rprop.logdens2,
  rprop.gen1,
  ...
)
}
\arguments{
\item{r}{Augmented variables in kernel matrix}

\item{delta}{Inclusion parameter for mixture components}

\item{lambda}{Variance component, tau*sigma^{-2}}

\item{y}{Outcome vector}

\item{X}{Covariate design matrix}

\item{beta}{Covariate parameter vector}

\item{sigsq.eps}{Variance component sigma^2}

\item{Vcomps}{Components of V_{lambda, Z, r} matrix, calculated with \code{\link{makeVcomps}}}

\item{Z}{Exposure design matrix}

\item{ztest}{Vector indicating which components of the mixture must remain in the model}

\item{data.comps}{Info for internal VComps.star calculation}

\item{control.params}{Control parameter list from \code{\link{kmbayes}}, can adjust MH proposal variance or prior values}

\item{rprop.gen2}{Proposal generation function for move type 2}

\item{rprop.logdens1}{Proposal log density function for move type 1}

\item{rprior.logdens}{Prior log density function for move type 1}

\item{rprior.logdens2}{Prior log density function for move type 2}

\item{rprop.logdens2}{Proposal log density function for move type 2}

\item{rprop.gen1}{Proposal generation function for move type 1}

\item{...}{Catches unneeded arguments}
}
\value{
MH step result with update r and delta parameters
}
\description{
Component-wise variable selection updates the r and delta parameters simultaneously by randomly selecting between two move options. See Bobb et al. (2015) supplement pp. 4-5 for details.
}
