% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{set_verbose_opts}
\alias{set_verbose_opts}
\title{Control bkmr printed output frequency and detail}
\usage{
set_verbose_opts(
  verbose_freq = NULL,
  verbose_show_ests = NULL,
  verbose_digits = NULL,
  tot_iter
)
}
\arguments{
\item{verbose_freq}{Frequency of output}

\item{verbose_show_ests}{Logical, show estimates or not}

\item{verbose_digits}{number of digits to print}

\item{tot_iter}{Number of iterations in function}
}
\value{
Returns list of options
}
\description{
Control bkmr printed output frequency and detail
}
