% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{sigsq.eps.update}
\alias{sigsq.eps.update}
\title{Update sigma^2 parameter in Gibbs step}
\usage{
sigsq.eps.update(y, X, beta, Vinv, a.eps = 0.001, b.eps = 0.001)
}
\arguments{
\item{y}{Outcome vector}

\item{X}{Covariate design matrix}

\item{beta}{Current draw of Beta parameters}

\item{Vinv}{Inverse of V_{lambda, Z, r}, calculated with \code{\link{makeVcomps}}}

\item{a.eps}{Prior shape parameter for Gamma distribution}

\item{b.eps}{Prior rate parameter for Gamma distribution}
}
\value{
Sample from full conditional posterior
}
\description{
See Bobb et al. (2015) supplement, page 4 for details.
}
