% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{TracePlot}
\alias{TracePlot}
\title{Generate Trace Plots for Parameters from a BKMR Fit}
\usage{
TracePlot(bkmrFit, par = "beta", comp = 1, subset = 1)
}
\arguments{
\item{bkmrFit}{An object of class \code{bkmrfit}, typically the result of the \code{kmbayes()} function containing MCMC samples.}

\item{par}{Character string specifying the parameter to plot the trace for.
Valid options are \code{"beta"}, \code{"lambda"}, \code{"sigsq.eps"}, \code{"h.hat"}, \code{"r"}, or \code{"delta"}.}

\item{comp}{Integer(s) or \code{"all"} indicating which component(s) of the parameter to plot. If \code{"all"}, trace plots are generated for all components.}

\item{subset}{An integer indicating which posterior sample subset to use. Defaults to 1.}
}
\value{
A list of ggplot2 objects (or a single plot) containing the trace plot(s) for the specified parameter/component.
}
\description{
Produces trace plots for posterior samples of parameters from a fitted Bayesian Kernel Machine Regression (BKMR) model.
This function is designed to work with objects created using the \code{kmbayes()} function.
}
\details{
This function extracts posterior samples for the chosen parameter and component(s), splits them by MCMC chain,
and generates trace plots using the helper function \code{genDiagnostic()}.
When multiple components are requested, the samples are reshaped into the format expected by \code{genDiagnostic()}.
}
\examples{
\dontrun{
fit <- kmbayes(...)  # assuming the fit has been run
TracePlot(fit, par = "beta", comp = 'all')         # Trace plot for all beta coefficiens
TracePlot(fit, par = "lambda", comp = 1)   # Trace plots for all lambda components
}

}
