% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divideConquer.R
\name{wasp_linProg}
\alias{wasp_linProg}
\title{Run Linear Program for calculating WASP}
\usage{
wasp_linProg(DList, atomMat, solver = "lpsolve")
}
\arguments{
\item{DList}{List of D matrices from WASP algorithm}

\item{atomMat}{Matrix of atoms from WASP algorithm}

\item{solver}{Linear program solver to use.}
}
\value{
Solution to WASP linear program
}
\description{
Linear program for calculating Wasserstein Posterior (WASP). See \code{\link{wasp_univariate}}.
}
