% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{ystar.update.noh}
\alias{ystar.update.noh}
\title{Update latent y^star values in Gibbs step without calculating h values (Probit model only)}
\usage{
ystar.update.noh(y, X, beta, Vinv, ystar)
}
\arguments{
\item{y}{Outcome vector}

\item{X}{Covariate design matrix}

\item{beta}{Current draw of Beta parameters}

\item{Vinv}{Current value of V_\{lambda, Z, r\}}

\item{ystar}{Current draws of ystar}
}
\value{
Vector of posterior draws from truncated normal distribution
}
\description{
See Bobb et al. (2018) for details. (Details sparse on this one)
}
