% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{ystar.update}
\alias{ystar.update}
\title{Update latent y^star values in Gibbs step (Probit model only)}
\usage{
ystar.update(y, X, beta, h)
}
\arguments{
\item{y}{Outcome vector}

\item{X}{Covariate design matrix}

\item{beta}{Current draw of Beta parameters}

\item{h}{Current draw of subject-specific health effects}
}
\value{
Vector of posterior draws from truncated normal distribution
}
\description{
See Bobb et al. (2018) for details.
}
