% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_gaussian}
\alias{bkmr_mcmc_gaussian}
\title{MCMC code to fit Gaussian BKMR Model}
\usage{
bkmr_mcmc_gaussian(
  y,
  Z,
  X,
  starting.values,
  nchains,
  iter,
  warmup,
  id,
  varsel,
  Znew,
  groups,
  rmethod,
  est.h,
  knots,
  control.params,
  verbose,
  missingX,
  ...
)
}
\arguments{
\item{y}{Outcome}
}
\description{
Code slightly modified from original bkmr package, Bobb 2015
}
