% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_logit}
\alias{bkmr_mcmc_logit}
\title{BKMR Sampling on Logit Model using Hamiltonian Monte Carlo}
\usage{
bkmr_mcmc_logit(
  y,
  Z,
  X,
  starting.values = list(rho = 2),
  nchains = 1,
  iter = 1000,
  warmup = 50,
  ...
)
}
\arguments{
\item{y}{Outcome}

\item{...}{Catch all to allow package flexibility}
}
\value{
STAN output from bkmr_logit fit
}
\description{
Fits bkmr_logit model using STAN. See file inst/stan/fit_logit.stan for STAN specifications.
}
