% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{bkmr_mcmc_poisson}
\alias{bkmr_mcmc_poisson}
\title{Redirect Poisson family using specified link function (default will be 'log')}
\usage{
bkmr_mcmc_poisson(link, ...)
}
\arguments{
\item{link}{character vector giving link function name. Currently implemented for 'log'}

\item{...}{All arguments passed through to sampler function}
}
\value{
Output from sampler function
}
\description{
Redirect Poisson family using specified link function (default will be 'log')
}
