% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmbayes.R
\name{kmbayes}
\alias{kmbayes}
\title{Fit Bayesian Kernel Machine Regression}
\usage{
kmbayes(
  y,
  Z,
  X,
  K = 1,
  iter = 1000,
  warmup = 50,
  nchains = 1,
  n_samps = iter,
  family = gaussian(),
  id = NULL,
  verbose = TRUE,
  Znew = NULL,
  starting.values = NULL,
  control.params = NULL,
  varsel = FALSE,
  groups = NULL,
  knots = NULL,
  ztest = NULL,
  rmethod = "varying",
  est.h = FALSE,
  WASPSolver = "lpsolve"
)
}
\arguments{
\item{y}{Outcome}

\item{WASPSolver}{}
}
\description{
Main function to fit BKMR using MCMC methods and divide-and-conquer approaches with WASP
}
\details{
TODO:
Make sure output works with other bkmr functions
Make sure all standard bkmr options still run
Deep dive on logit and poisson sampling functions
}
