% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_update_functions.R
\name{r.update}
\alias{r.update}
\title{Sample new r values for exposures fixed to be in the model}
\usage{
r.update(
  r,
  theta,
  whichcomp,
  delta,
  lambda,
  y,
  X,
  beta,
  sigsq.eps,
  Vcomps,
  Z,
  data.comps,
  control.params,
  rprop.gen,
  rprop.logdens,
  rprior.logdens,
  M = 1,
  ...
)
}
\arguments{
\item{r}{Augmented variables in kernel matrix}

\item{theta}{Function smoothness parameter in prior for r}

\item{whichcomp}{Integer vector indicating which r values should be updated}

\item{delta}{Inclusion parameter for mixture components}

\item{lambda}{Variance component, tau*sigma^\{-2\}}

\item{y}{Outcome vector}

\item{X}{Covariate design matrix}

\item{beta}{Covariate parameter vector}

\item{sigsq.eps}{Variance component sigma^2}

\item{Vcomps}{Components of V_\{lambda, Z, r\} matrix, calculated with \code{\link{makeVcomps}}}

\item{Z}{Exposure design matrix}

\item{data.comps}{Info for internal VComps.star calculation}

\item{control.params}{Control parameter list from \code{\link{kmbayes}}, can adjust MH proposal variance or prior values}

\item{rprop.gen}{Function for generating new r proposal value}

\item{rprop.logdens}{Function to calculate log density for proposed r value}

\item{rprior.logdens}{Function to calculate log density for proposed r under prior}

\item{M}{Stochastic approximation constant. If StA is requested, M = K, otherwise M = 1.}

\item{...}{Catches unused arguments}
}
\value{
MH Step result containing updated r values
}
\description{
Sample new r values for exposures fixed to be in the model
}
